requirements.txt
torch==2.1.2
transformers==4.39.3
diffusers==0.27.2
safetensors==0.5.2
Pillow==10.1.0
accelerate==0.24.1
numpy<2
huggingface_hub==0.21.4
